home-only template: exclude NixOS modules (#151)
This PR fixes #147.

The issue was with the attribute set returned by `filters` (in
`/modules/flake/template.nix`). The `nixosOnly` filter excludes paths
containing `modules/nixos/common` as an infix. However, the home
template only considers paths that the `nixosOnly` or `darwinOnly`
filters return false for. As a result `modules/nixos/common` was being
included in the home template.

This was fixed by adding a new filter to the attribute set,
`homeFilter`.

Co-authored-by: Sridhar Ratnakumar <3998+srid@users.noreply.github.com>